+++ pyro/params/__init__.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import, division, print_function
 
-from pyro.params.param_store import _MODULE_NAMESPACE_DIVIDER, _PYRO_PARAM_STORE  # noqa: F401
 from pyro.params.param_store import module_from_param_with_module_name, param_with_module_name, user_param_name
 
 __all__ = [
