+++ pyro/params/param_store.py
@@ -175,3 +175,21 @@ class ParamStoreDict(object):
         with open(filename, "rb") as input_file:
             state = torch.load(input_file)
         self.set_state(state)
+
+
+# used to create fully-formed param names, e.g. mymodule$$$mysubmodule.weight
+_MODULE_NAMESPACE_DIVIDER = "$$$"
+
+
+def param_with_module_name(pyro_name, param_name):
+    return _MODULE_NAMESPACE_DIVIDER.join([pyro_name, param_name])
+
+
+def module_from_param_with_module_name(param_name):
+    return param_name.split(_MODULE_NAMESPACE_DIVIDER)[0]
+
+
+def user_param_name(param_name):
+    if _MODULE_NAMESPACE_DIVIDER in param_name:
+        return param_name.split(_MODULE_NAMESPACE_DIVIDER)[1]
+    return param_name
