+++ pyro/primitives.py
@@ -14,6 +14,7 @@ import pyro.distributions as dist
 import pyro.infer as infer
 import pyro.poutine as poutine
 from pyro.distributions.distribution import Distribution
+from pyro.params import param_with_module_name
 from pyro.poutine.runtime import _DIM_ALLOCATOR, _PYRO_PARAM_STORE, _MODULE_NAMESPACE_DIVIDER, am_i_wrapped, apply_stack
 from pyro.util import deep_getattr, set_rng_seed  # noqa: F401
 
@@ -437,17 +438,3 @@ def validation_enabled(is_validate=True):
     finally:
         dist.enable_validation(distribution_validation_status)
         infer.enable_validation(infer_validation_status)
-
-
-def param_with_module_name(pyro_name, param_name):
-    return _MODULE_NAMESPACE_DIVIDER.join([pyro_name, param_name])
-
-
-def module_from_param_with_module_name(param_name):
-    return param_name.split(_MODULE_NAMESPACE_DIVIDER)[0]
-
-
-def user_param_name(param_name):
-    if _MODULE_NAMESPACE_DIVIDER in param_name:
-        return param_name.split(_MODULE_NAMESPACE_DIVIDER)[1]
-    return param_name
